fix(ui): moving the unified state fetching into its own component
made more sense than all of the copy pasta that was being slung around
in the previous commit. Tests should still live on the form.
