+++ ui/src/variables/components/VariableForm.test.tsx
@@ -60,32 +60,23 @@ describe('VariableForm', () => {
   })
 
   describe('type selector', () => {
-    it('should not update state if the type was not changed', () => {
+    it('should not tell the the parrent if the type was not changed', () => {
       const {wrapper, actions} = setup()
 
       const defaultType = 'query'
-      const stateStub = jest.fn()
-
-      wrapper.instance().setState = stateStub
 
       // this way of accessing the instance function
       // bypasses typescript's private function check
       wrapper.instance()['handleChangeType'](defaultType)
-      expect(stateStub.mock.calls.length).toBe(0)
       expect(actions.type.mock.calls.length).toBe(0)
     })
 
-    it('should update state if the type was changed', () => {
+    it('should update tell the the parrent if type was changed', () => {
       const {wrapper, actions} = setup()
-
-      const stateStub = jest.fn()
-
-      wrapper.instance().setState = stateStub
 
       // this way of accessing the instance function
       // bypasses typescript's private function check
       wrapper.instance()['handleChangeType']('map')
-      expect(stateStub.mock.calls.length).toBe(1)
       expect(actions.type.mock.calls.length).toBe(1)
       expect(actions.type.mock.calls[0][0]).toBe('map')
     })
